+++ addons/ui_react/scripts/internal/react/ui_react_base_button_reactive.gd
@@ -24,6 +24,7 @@ func on_enter_tree() -> void:
 
 
 func on_exit_tree() -> void:
+	disconnect_all_states()
 	_TxnSession.unregister_host(_host)